#1007 Split resolver interface and implementation, clean up runtime implementations
